+++ packages/account-wasm/src/lib.rs
@@ -11,7 +11,8 @@ use account_sdk::account::outside_execution::{OutsideExecutionAccount, OutsideEx
 use account_sdk::account::session::SessionAccount;
 use account_sdk::account::AccountHashAndCallsSigner;
 use account_sdk::constants::{Version, CONTROLLERS};
-use account_sdk::controller::{Controller, ControllerError};
+use account_sdk::controller::Controller;
+use account_sdk::errors::ControllerError;
 use account_sdk::provider::{CartridgeJsonRpcProvider, CartridgeProvider};
 use account_sdk::signers::webauthn::{CredentialID, WebauthnSigner};
 use account_sdk::signers::{HashSigner, Signer};
